# Conflicts:
#	src/main/java/com/ccorder/ordersystem/controller/FoodController.java
#	src/main/java/com/ccorder/ordersystem/controller/UserInoController.java

+++ src/main/java/com/ccorder/ordersystem/controller/FoodController.java
@@ -63,14 +63,13 @@ public class FoodController {
     @ApiOperation(value = "获取客户端食物列表")
     @GetMapping("getUserFoodList")
     @ResponseBody
-    public Object getUserFoodList() {
+    public Object getUserFoodList(){
 
         @Data
-        class kindFoods {
+        class kindFoods{
             SysDict type;
             UserFood[] typeFoodList;
-
-            kindFoods(SysDict type, UserFood[] userFoods) {
+            kindFoods(SysDict type,UserFood[] userFoods){
                 this.type = type;
                 this.typeFoodList = userFoods;
             }
@@ -78,30 +77,30 @@ public class FoodController {
         List<String> foodtypesID = new ArrayList<>();
         List<kindFoods> foodsList = new ArrayList<>();
 
-        try {
-            List<Map<String, Object>> types = foodService.getAllType();
-            for (int i = 0; i < types.size(); i++) {
+        try{
+            List<Map<String,Object>> types = foodService.getAllType();
+            for (int i = 0 ; i < types.size();i++){
                 foodtypesID.add(types.get(i).get("foodType").toString());
             }
-            for (int i = 0; i < foodtypesID.size(); i++) {
+            for (int i = 0; i< foodtypesID.size();i++){
                 System.out.println(foodtypesID.get(i));
                 UserFood[] userFoods = foodService.getFoodsByType(foodtypesID.get(i));
-                foodsList.add(new kindFoods(sysDictService.selectByPrimaryKey(foodtypesID.get(i)), userFoods));
+                foodsList.add(new kindFoods(sysDictService.selectByPrimaryKey(foodtypesID.get(i)),userFoods));
             }
 
-            return new AjaxMessage().Set(MsgType.Success, "获取食品列表成功", foodsList);
-        } catch (Exception e) {
+            return new AjaxMessage().Set(MsgType.Success,"获取食品列表成功", foodsList);
+        }catch (Exception e){
             e.printStackTrace();
         }
-        return new AjaxMessage().Set(MsgType.Error, "获取食品列表失败");
+        return new AjaxMessage().Set(MsgType.Error,"获取食品列表失败");
     }
 
     @PostMapping("/addNewUserFood")
     @ResponseBody
     public Object addNewFood(
             @RequestBody
-                    UserFood userFood
-    ) {
+            UserFood userFood
+    ){
         return new AjaxMessage().Set(MsgType.Success, "Just For Model");
     }
 
@@ -109,27 +108,18 @@ public class FoodController {
     @PostMapping("/addNewType")
     @ResponseBody
     public Object addNewType(
-            @ApiParam(name = "newFoodType", value = "新的Food种类(中文)", required = true, type = "String")
-            @RequestParam
-                    String newFoodType,
-            @ApiParam(name = "sortNum", value = "类别排序数字", required = true, type = "Integer")
-            @RequestParam
-                    Integer sortNum
+            @ApiParam(name = "newFoodType", value = "新的Food种类(中文)", required = true,type = "SysDict")
+            @RequestBody
+                    SysDict newFoodType
     ) {
-
-        /*获取食品类别的dict_type_id*/
-        String dictTypeId = dictMapper.selectByNameEnAndStatus("foodType", 0);
-        SysDict newSysDict = new SysDict();
-        newSysDict.setId(UUID.randomUUID().toString());
-        newSysDict.setNameCn(newFoodType);
-        newSysDict.setSort(sortNum);
-
-        Date dateNow = new Date();
-        newSysDict.setCreateDate(dateNow);
-        newSysDict.setModifyDate(dateNow);
-
-        try {
-            dictMapper.insertSelective(newSysDict);
+       Date nowDate=new Date();
+       newFoodType.setId(UUID.randomUUID().toString());
+       newFoodType.setCreateDate(nowDate);
+       newFoodType.setModifyDate(nowDate);
+       newFoodType.setDictType(newFoodType.getDictType());
+       newFoodType.setSort(sysDictService.countAllRecords()+1);
+       try {
+            sysDictService.insertSelective(newFoodType);
             return new AjaxMessage().Set(MsgType.Success, "添加食品分类成功");
         } catch (Exception e) {
             e.printStackTrace();
@@ -156,7 +146,12 @@ public class FoodController {
         if (storeUser != null) {
             try {
                 /*food表插入记录*/
+                Date tmpDate = new Date();
                 newFood.setId(UUID.randomUUID().toString());
+                newFood.setCreateUserId(newFood.getCreateUserId());
+                newFood.setModifyUserId(newFood.getModifyUserId());
+                newFood.setCreateDate(tmpDate);
+                newFood.setModifyDate(tmpDate);
                 foodService.addNewFood(newFood);
 
                 /*添加关联记录*/
@@ -166,7 +161,7 @@ public class FoodController {
                 userFoodMap.setFoodId(newFood.getId());
                 userFoodMap.setCreateUserId(storeUser.getId());
                 userFoodMap.setModifyUserId(storeUser.getId());
-                Date tmpDate = new Date();
+
                 userFoodMap.setCreateDate(tmpDate);
                 userFoodMap.setModifyDate(tmpDate);
                 mapUserFoodMapper.insertSelective(userFoodMap);
@@ -183,17 +178,20 @@ public class FoodController {
     @ResponseBody
     public Object deleteOneFood(
             @ApiParam(name = "foodId", value = "需要删除的食品的id", required = true, type = "String")
-            @RequestParam
-             String foodId
+            @RequestParam(value = "foodId")
+                    String foodId
     ) {
         try {
+            if(foodMapper.selectByFoodId(foodId)==null){
+                return new AjaxMessage().Set(MsgType.Success,"该食物不存在");
+            }
             foodMapper.deleteByPrimaryKey(foodId);
-            return new AjaxMessage().Set(MsgType.Success, "成功删除食物");
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return new AjaxMessage().Set(MsgType.Success, "删除食物失败");
+            return new AjaxMessage().Set(MsgType.Success,"成功删除食物");
+    }catch (Exception e){
+        e.printStackTrace();
     }
+        return new AjaxMessage().Set(MsgType.Success,"删除食物失败");
+}
 
     @ApiOperation(value = "商家更新某一个食品的信息")
     @PostMapping("/updateOneFood")
@@ -205,15 +203,15 @@ public class FoodController {
     ) {
         try {
             //先判断food是否存在
-            if (foodMapper.selectByFoodId(food.getId()) == null) {
-                return new AjaxMessage().Set(MsgType.Success, "不存在此食物");
+            if(foodMapper.selectByFoodId(food.getId())==null){
+                return new AjaxMessage().Set(MsgType.Success,"不存在此食物");
             }
-            foodMapper.updateByPrimaryKey(food);
-            return new AjaxMessage().Set(MsgType.Success, "更新食物成功");
-        } catch (Exception e) {
+            foodMapper.updateByPrimaryKeySelective(food);
+            return new AjaxMessage().Set(MsgType.Success,"更新食物成功");
+        }catch (Exception e){
             e.printStackTrace();
         }
-        return new AjaxMessage().Set(MsgType.Success, "更新食物失败");
+        return new AjaxMessage().Set(MsgType.Success,"更新食物失败");
     }
 
     @ApiOperation(value = "商家删除某一类的食物")
@@ -221,15 +219,15 @@ public class FoodController {
     @ResponseBody
     public Object deleteOneTypeFood(
             @ApiParam(name = "foodType", value = "需要更新的食品", required = true, type = "String")
-            @RequestBody
+            @RequestParam(value = "foodType")
                     String foodType
     ) {
         try {
             foodMapper.deleteByFoodType(foodType);
-            return new AjaxMessage().Set(MsgType.Success, "删除此类食物成功");
-        } catch (Exception e) {
+            return new AjaxMessage().Set(MsgType.Success,"删除此类食物成功");
+        }catch (Exception e){
             e.printStackTrace();
         }
-        return new AjaxMessage().Set(MsgType.Success, "删除此类食物失败");
+        return new AjaxMessage().Set(MsgType.Success,"删除此类食物失败");
     }
 }
